NutriScan/requirements.txt
fastapi
uvicorn
pydantic
email-validator
requests
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
